third_party/blink/web_tests/html5lib/generated/run-tests18-data-expected.txt

CONSOLE ERROR: line 1: Uncaught SyntaxError: Unexpected token '<'
CONSOLE ERROR: line 1: Uncaught SyntaxError: Unexpected token '<'
CONSOLE ERROR: line 1: Uncaught SyntaxError: Unexpected token '<'
CONSOLE ERROR: line 1: Uncaught SyntaxError: Unexpected token '<'
../resources/tests18.dat: PASS
